Further refactor `http` into a module
Note, this uses the newer sibling mod style instead of `mod.rs`, which is inconsistent with the rest of the project but considered more idiomatic.
