+++ apollo-federation/src/sources/connect/validation/http.rs
@@ -1,0 +1,2 @@
+pub(super) mod headers;
+pub(super) mod method;
