+++ apollo-federation/src/sources/connect/validation/mod.rs
@@ -17,8 +17,7 @@
 mod coordinates;
 mod entity;
 mod extended_type;
-mod http_headers;
-mod http_method;
+mod http;
 mod selection;
 mod source_name;
 
@@ -40,8 +39,6 @@ use apollo_compiler::Schema;
 use coordinates::source_base_url_argument_coordinate;
 use coordinates::source_http_argument_coordinate;
 use extended_type::validate_extended_type;
-use http_headers::get_http_headers_arg;
-use http_headers::validate_headers_arg;
 use itertools::Itertools;
 use source_name::SourceName;
 use url::Url;
@@ -53,6 +50,7 @@ use crate::sources::connect::spec::schema::HTTP_ARGUMENT_NAME;
 use crate::sources::connect::spec::schema::SOURCE_BASE_URL_ARGUMENT_NAME;
 use crate::sources::connect::spec::schema::SOURCE_DIRECTIVE_NAME_IN_SPEC;
 use crate::sources::connect::spec::schema::SOURCE_NAME_ARGUMENT_NAME;
+use crate::sources::connect::validation::http::headers;
 use crate::sources::connect::ConnectSpecDefinition;
 use crate::subgraph::spec::CONTEXT_DIRECTIVE_NAME;
 use crate::subgraph::spec::EXTERNAL_DIRECTIVE_NAME;
@@ -307,11 +305,11 @@ fn validate_source(directive: &Component<Directive>, sources: &SourceMap) -> Sou
             }
         }
 
-        // Validate headers argument
-        if let Some(headers) = get_http_headers_arg(http_arg) {
-            let header_errors = validate_headers_arg(&directive.name, headers, sources, None, None);
-            errors.extend(header_errors);
-        }
+        errors.extend(
+            headers::validate_arg(http_arg, &directive.name, sources, None, None)
+                .into_iter()
+                .flatten(),
+        );
     } else {
         errors.push(Message {
             code: Code::GraphQLError,
